requirements.txt
python-telegram-bot[job-queue]==21.7
aiohttp==3.9.1
aiosqlite==0.19.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
cryptography==41.0.7
Jinja2==3.1.2